Split GraphQL wire layers into wire submodules
Extract the codegen derives, envelope structs, `CursorVariables`
impls, and domain-type definitions from `src/review_threads.rs`
(505 lines) and `src/reviews.rs` (425 lines) into
`src/review_threads/wire.rs` and `src/reviews/wire.rs`, bringing every
product source back under the 400-line limit. Public paths are
preserved through `pub use wire::…` re-exports, so no consumer
changes.

Pure refactor: no behaviour or signature changes.
`src/review_threads/tests.rs` (655 lines, tests-only) already exceeded
the limit at HEAD and is left for a follow-up.
